Database connect

+++ PaymentGateway.PublishedLanguage/Events/DepositCreated.cs
@@ -5,10 +5,10 @@ namespace PaymentGateway.PublishedLanguage.Events
     public class DepositCreated : INotification
     {
         public string Iban { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Currency { get; set; }
 
-        public DepositCreated(string iban, double amount, string currency)
+        public DepositCreated(string iban, decimal amount, string currency)
         {
             this.Iban = iban;
             this.Amount = amount;
