+++ PaymentGateway.PublishedLanguage/Events/ProductCreated.cs
@@ -6,11 +6,11 @@ namespace PaymentGateway.PublishedLanguage.Events
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
         public string Currency { get; set; }
         public int Limit { get; set; }
 
-        public ProductCreated(int id, string name, double value, string currency, int limit)
+        public ProductCreated(int id, string name, decimal value, string currency, int limit)
         {
             Id = id;
             Name = name;
